Refactor MPI utility functions
