+++ source/source_esolver/esolver_ks_lcao_tddft.h
@@ -2,51 +2,13 @@
 #define ESOLVER_KS_LCAO_TDDFT_H
 #include "esolver_ks.h"
 #include "esolver_ks_lcao.h"
-#include "source_base/module_container/ATen/core/tensor.h"   // ct::Tensor
-#include "source_base/module_external/scalapack_connector.h" // Cpxgemr2d
+#include "source_base/module_container/ATen/core/tensor.h" // ct::Tensor
+#include "source_lcao/module_rt/gather_mat.h"              // MPI gathering and distributing functions
 #include "source_lcao/module_rt/td_info.h"
 #include "source_lcao/module_rt/velocity_op.h"
 
 namespace ModuleESolver
 {
-//------------------------ MPI gathering and distributing functions ------------------------//
-// This struct is used for collecting matrices from all processes to root process
-template <typename T>
-struct Matrix_g
-{
-    std::shared_ptr<T> p;
-    size_t row;
-    size_t col;
-    std::shared_ptr<int> desc;
-};
-
-// Collect matrices from all processes to root process
-template <typename T>
-void gatherMatrix(const int myid, const int root_proc, const hamilt::MatrixBlock<T>& mat_l, Matrix_g<T>& mat_g)
-{
-    const int* desca = mat_l.desc; // Obtain the descriptor of the local matrix
-    int ctxt = desca[1];           // BLACS context
-    int nrows = desca[2];          // Global matrix row number
-    int ncols = desca[3];          // Global matrix column number
-
-    if (myid == root_proc)
-    {
-        mat_g.p.reset(new T[nrows * ncols]); // No need to delete[] since it is a shared_ptr
-    }
-    else
-    {
-        mat_g.p.reset(new T[nrows * ncols]); // Placeholder for non-root processes
-    }
-
-    // Set the descriptor of the global matrix
-    mat_g.desc.reset(new int[9]{1, ctxt, nrows, ncols, nrows, ncols, 0, 0, nrows});
-    mat_g.row = nrows;
-    mat_g.col = ncols;
-
-    // Call the Cpxgemr2d function in ScaLAPACK to collect the matrix data
-    Cpxgemr2d(nrows, ncols, mat_l.p, 1, 1, const_cast<int*>(desca), mat_g.p.get(), 1, 1, mat_g.desc.get(), ctxt);
-}
-//------------------------ MPI gathering and distributing functions ------------------------//
 
 template <typename TR, typename Device = base_device::DEVICE_CPU>
 class ESolver_KS_LCAO_TDDFT : public ESolver_KS_LCAO<std::complex<double>, TR>
