+++ source/source_lcao/module_rt/evolve_psi.cpp
@@ -176,15 +176,15 @@ void evolve_psi_tensor(const int nband,
         MPI_Comm_rank(MPI_COMM_WORLD, &myid);
         MPI_Comm_size(MPI_COMM_WORLD, &num_procs);
 
-        ModuleESolver::Matrix_g<std::complex<double>> h_mat_g, s_mat_g; // Global matrix structure
+        module_rt::Matrix_g<std::complex<double>> h_mat_g, s_mat_g; // Global matrix structure
 
         // Collect H matrix
-        ModuleESolver::gatherMatrix(myid, 0, h_mat, h_mat_g);
+        module_rt::gatherMatrix(myid, 0, h_mat, h_mat_g);
         syncmem_complex_h2d_op()(Htmp.data<std::complex<double>>(), h_mat_g.p.get(), len_HS);
         syncmem_complex_h2d_op()(Hold.data<std::complex<double>>(), h_mat_g.p.get(), len_HS);
 
         // Collect S matrix
-        ModuleESolver::gatherMatrix(myid, 0, s_mat, s_mat_g);
+        module_rt::gatherMatrix(myid, 0, s_mat, s_mat_g);
         syncmem_complex_h2d_op()(Stmp.data<std::complex<double>>(), s_mat_g.p.get(), len_HS);
     }
     else
